events: Move RoomMessageEventContentWithoutRelation into separate module
